+++ js/interactions/index.js
@@ -1,0 +1,9 @@
+/**
+ * WOW3 Interactions Index
+ * Central export point for all interaction handler classes
+ */
+
+export { DragHandler } from './DragHandler.js';
+export { ResizeHandler } from './ResizeHandler.js';
+export { RotateHandler } from './RotateHandler.js';
+export { AlignmentGuides } from './AlignmentGuides.js';